fix: correct test for checking if public paginated travels are returned due to wrong precision
The expected response data contained all the public travels, but since our pagination only allows 10 by page, whenever the public travels amount is greater than 10, the test fails: not all public travels are returned in the response, just the first 10. Now we estipulate exactly 10 travels to be created in the test, 5 non-public and 5 public, and iterate over all the public ones to see if they're present in the response (if the application is ok they'll ever be present: 5 is always less than 10, so they're in the same page, thus inside the json response).